fix(web): exact integers on the wire, validated routes, no symlink escape
JsonWriter::num sent every integer through %.10g, so a byte count rounded
to 1.2e+10 past 10 GB and so did the drop counters; integral types now
print their digits. add_http_route validated nothing, so an empty or "/"
prefix shadowed the client bundle, and it wrote _routes unlocked while
HTTP threads read them. fields_of accepted trailing junk after the closing
brace. safe_name clears a name but not where it resolves, so the
recordings route now checks the file lands inside --record-dir, and app
routes get the same Host check as the WebSocket upgrade.

+++ desktop_examples/websdr/web_demo.cpp
@@ -53,8 +53,8 @@ int main(int argc, char** argv) {
     srv.set_state(state_json(gen, freq, mode, gain, source));
     srv.add_http_route("/health", [&srv, &o](const std::string&, const std::string&) {
         web::JsonWriter w;
-        w.begin_obj().key("version").str(o.version).key("uptime_s").num(static_cast<double>(srv.uptime_seconds()))
-         .key("clients").num(static_cast<double>(srv.client_count())).key("source").str("sim").end();
+        w.begin_obj().key("version").str(o.version).key("uptime_s").num(srv.uptime_seconds())
+         .key("clients").num(srv.client_count()).key("source").str("sim").end();
         return web::HttpReply{200, w.out, "application/json"};
     });
     srv.start();
